GMimeCertificate: add User ID
GPGME reports full User IDs as well as name and e-mail address.  We
should also be capable of exposing the full User ID, as not every User
ID will be formed from exactly a name and an E-mail address.  This
changeset makes it possible to export the User ID from GMime.

Note that this is an API change and an ABI extension, but it's one
that has no effect on anyone who uses GMime in the standard way (with
constructors, destructors, and accessor methods, rather than static
allocation of GMime objects on the stack or direct manipulation of the
underlying struct).

This API extension is comparable to the addition of
retrieve_session_key to the GMimeGpgContext last year (in the
transition from 2.6.20 to 2.6.21), so hopefully it's not considered a
breaking change.

fixup add user_id to GMimeCertificate