GMimeCertificate: expose identity validity
The GMimeCertificate currently doesn't expose any indication of
whether the user should believe the identity assertions contained in
the certificate.  This changeset provides a mechanism for exposing
that information.

A subsequent revision will actually set that value so that it is
visible to the user of the GMime library.

Note that this is an API change and an ABI extension, but it's one
that has no effect on anyone who uses GMime in the standard way (with
constructors, destructors, and accessor functions, rather than static
allocation of GMime objects on the stack or direct manipulation of the
underlying struct).

This API extension is comparable to the addition of
retrieve_session_key to the GMimeGpgContext last year (in the
transition from 2.6.20 to 2.6.21) and the addition of user_id to
GMimeCertificate (in this series), so hopefully it's not considered a
breaking change.