Autocrypt: introduce GMimeAutocryptHeaderList
MUAs that implement Autocrypt are likely to deal with batches of
Autocrypt headers associated with an e-mail.

In the weird corner case where an e-mail has multiple addresses listed
in the From: line, for example, it's possible for there to be multiple
Autocrypt headers that relate to the different senders of the e-mail.

Likewise, gossiped keys (simplified key introductions usually sent
with encrypted mail), we'll also need to deal with a set of Autocrypt
headers.

This abstraction will be useful for MUA implementors when scanning
e-mail that has multiple addresses either from or to.