Autocrypt: restructure g_mime_message_get_autocrypt_headers
Getting gossiped Autocrypt headers (introductions from a third party)
is similar to getting Autocrypt headers about the sender, but with a
few minor differences.  We abstract out the work of picking up the
headers here, to make it easier to implement retrieving gossiped
headers, which will happen in a subsequent patch.

Note that MUAs that handle gossiped headers will need distinct logic
for dealing with gossiped headers instead of "From:" Autocrypt
headers, so it makes sense that they would expect a different API call
to retrieve the gossiped headers.