Fixed header parsing logic when dealing with input in DOS format
When the input is in DOS format, the <CR> of the blank line
terminating the headers would be appended to priv->headerbuf
and then, after breaking out of the loop, it would be parsed
as a header (which, of course, would fail and so be set as
priv->preheader and later message->marker).

This would then cause writing out of message/rfc822 parts
to be written with a marker of "\r" (+ the newline sequence
specified by the GMimeFormatOptions).

Fixes issue #23