Implement basic error reporting for the GMime parser (#26)
Add an error reporting facility for the GMime parser.  Basically, it extends GMimeParserOptions
by a callback which is called whenever the parser detects a “fishy” construct.  The callback
receives the offset within the message, an error code and (if applicable) an additional string,
plus a user-defined data pointer.

The errors reported are at the moment:
- conflicting Content-* headers (i.e. different Content-* headers for the same part)
- conflicting header parameters (ditto)
- non-conflicting Content-* headers (i.e. exactly the same header is repeated)
- non-conflicting header parameters (ditto)
- unencoded 8-bit characters in headers
- invalid content-type headers and completely broken headers
- malformed multipart/* and malformed messages
- truncated messages
- multipart without boundary