Autocrypt update to spec Level 1.0.0rc1  (#27)
* Drop Autocrypt type value

At the Autocrypt meeting this weekend, we determined that we can
completely drop the explicit "type=" attribute for the Autocrypt
header.

This change removes the type from the gmime API, simplifying it.
Hopefully it is ok because there has been no release with this API.

* Move _get_autocrypt_headers to gmime-object.c

This prepares for a shift in API so that we can get gossip headers
from the child of a GMimeMultipartEncrypted object instead of from a
GMimeMessage object.

* Only a single Autocrypt header can be returned from a GMimeMessage

As the Autocrypt spec reached its first release, we've clarified that
only a single address is ever valid from an e-mail message with
respect to Autocrypt.

See https://github.com/autocrypt/autocrypt/issues/138 and
https://github.com/autocrypt/autocrypt/pull/144 for discussion and
Autocrypt spec commit 45df7b1ea8fe6e83bbe8abae09dc495e4f7ab784 for
the clarification.

This commit adjusts the GMime interface to match the spec.

* autocrypt: add several more test messages

* Autocrypt: introduce gm_message_get_autocrypt_headers_from_inner_part

Autocrypt Level 1 has clarified that gossip headers must only be found
within the encrypted layer.  If a message has no encryption layer,
there should be no gossip processing.

See discussion at https://github.com/autocrypt/autocrypt/issues/135
and Autocrypt commit 1636162ee10df2d61d614e3903a5ff35cb953a8c for the
specific changes in the spec.

This changeset updates the retrieval of gossip headers so anyone who
has decrypted the message can retrieve them from the inner part, once
decrypted.  This is an API change against the currently unreleased
Autocrypt API.

* Autocrypt: added OpenPGP secret key to test in preparation for actual decryption.

* Autocrypt: fetch gossip headers from encrypted part.

This changeset adds a convenience function for any implementer to
extract key gossip information from an encrypted message, without
having to do the decryption manually.

We also add a new element to the test suite to exercise it and make
sure it works as expected.