Autocrypt: add g_mime_object_add_autocrypt_gossip_headers()
Gossip headers should be added to an inner MIME part before
encryption.  This function makes it easy to add these headers to a
GMimeObjectfrom any available GMimeAutocryptHeaderList.

We add a chunk of test suite to exercise this code as well as
round-tripping it through encryption/decryption, as is expected to be
the standard use case.