Expose signature creation time and expiration time as guint64
This is a very conservative change that attempts to address #68.

It avoids changing the memory layout of the GMimeSignature object,
since it might accessed directly by a user who doesn't use the
function interface.

It takes a different approach to indicating a failure than the
established (time_t)-1 mechanism, to encourage safer use.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>