build: avoid useless configure-time checks
Checking for the sizeof various types and defining preprocessor macros
for them, is pointless, since the *only* effect these had was defining
said macros... and the code never did anything with them.

This speeds up running configure and is less confusing to readers. It's
also less confusing to porters, as evidenced by the VS projects which
contain out of date copies of the config.h files that are missing some
of these unused checks.