Make gir follow the spec.
Without this change, the generated gir doesn't follow the spec and some
tools fail to parse the gir. This isn't a problem overall because these
are skipped because they are marked as not introspectable. But for some
languages (rust), the introspectable tag is more of a suggestion and
(skip)-ed objects and functions are still parsed. Not having a name for
the type creates an error which in part breaks the gir generation tool.